fix repertorios filters. Fix select productoras.
