+++ src/components/ModalProvider/ModalProvider.tsx
@@ -104,8 +104,6 @@ const ModalProvider: FC<ModalProvderProps> = ({ children }) => {
     switch (modalData.type) {
       case ModalNames.COMPLETE_REGISTRATION:
         return <EndRegisterUserModal />;
-      case ModalNames.CHANGE_PRODUCER:
-        return <ChangeProducerModal onCloseModal={onCloseModal} />;
       case ModalNames.ADD_TERRITORIALITY:
         return <AddTerritorialityModal onCloseModal={onCloseModal} />;
       case ModalNames.SEARCH_CONFLICTS_FILTERS:
@@ -225,80 +223,6 @@ const EndRegisterUserModal: FC = () => {
       </p>
 
       <CustomButton>Completar Registro</CustomButton>
-    </div>
-  );
-};
-
-const ChangeProducerModal: FC<{ onCloseModal: () => void }> = ({
-  onCloseModal,
-}) => {
-  const dispatch = useAppDispatch();
-  const authData = useAppSelector((state) => state.auth);
-
-  const selectProductora = async (element: {
-    id: string;
-    productora: string;
-  }) => {
-    try {
-      await selectProductionCompany(element.id);
-      const company = await getProducerById(element.id);
-      const productionCompany = { ...element, cuit_cuil: company.cuit_cuil };
-      if (window && window.localStorage) {
-        localStorage.setItem("company", JSON.stringify(productionCompany));
-      }
-      dispatch(
-        setAuthData({ ...authData, productoraActiva: productionCompany })
-      );
-    } catch (error) {
-      console.log(error);
-    } finally {
-      onCloseModal();
-    }
-  };
-
-  useEffect(() => {
-    if (
-      authData.productoras &&
-      authData.id_usuario &&
-      !authData.productoras[0].id
-    ) {
-      onCloseModal();
-    }
-  }, [authData]);
-  return (
-    <div className="relative bg-white h-[13rem] w-[30rem] mb-[6rem] rounded-[2rem] gap-[0.5rem] flex flex-col justify-center items-center">
-      <button onClick={onCloseModal} className="absolute top-[5%] right-[5%]">
-        <IoClose size={25} color="black" />
-      </button>
-
-      <p className="text-black font-bold text-[1.2rem] text-center w-[95%]">
-        Selecciona una productora.
-      </p>
-      {authData.productoras && authData.productoras.length > 0
-        ? authData.productoras.map((element, index) => (
-            <div
-              key={index}
-              className="w-[100%] cursor-pointer"
-              onClick={() => selectProductora(element)}
-            >
-              <p className="text-center text-black hover:bg-[#d8d8d8]">
-                {element.productora}
-              </p>
-            </div>
-          ))
-        : null}
-      {/* <div
-        onClick={() => handleChangeProductionModal("SONY MUSIC")}
-        className="w-[100%] cursor-pointer"
-      >
-        <p className="text-center text-black hover:bg-[#d8d8d8]">Sony Music</p>
-      </div>
-      <div
-        onClick={() => handleChangeProductionModal("GOLDSTEIN")}
-        className="w-[100%] cursor-pointer"
-      >
-        <p className="text-center text-black hover:bg-[#d8d8d8]">Goldstein</p>
-      </div> */}
     </div>
   );
 };
